Refactor: Update UserDAO namespace to App\DAO\Auth
